Fix for Desktop app
Seems the latest changes broke the desktop app as it uses
phonegap-connect and only copies the `www` folder, not the entire root.

This moves `config.xml` and `res` back into `www` (as well as lightening
the file size on `res` by removing the gradient in the splash screens).

See https://github.com/phonegap/phonegap-app-desktop/issues/584 etc