An example of using `semistandard`
Includes

1. semistandard - https://github.com/Flet/semistandard
2. snazzy (pretty formatting of errors, not 100% needed)
3. npm script to run semistandard as a test

Also includes the changes this particular repo would need to pass

Pros:

Bolts-included full set of reasonable standards.

Cons:

Not meant to be tweaked for personal preference. Take it or leave it
approach.

A more tweakable way of implementing semistandard as a "base" would be
an eslint config such as https://github.com/Flet/eslint-config-semistandard